+++ node.gyp
@@ -157,7 +157,6 @@
       'deps/node-inspect/lib/internal/inspect_client.js',
       'deps/node-inspect/lib/internal/inspect_repl.js',
       'deps/acorn/dist/acorn.js',
-      'deps/acorn/dist/acorn_loose.js',
       'deps/acorn/dist/walk.js',
     ],
     'conditions': [
