+++ test/parallel/test-repl-tab-complete.js
@@ -205,7 +205,7 @@ testMe.complete('toSt', common.mustCall(function(error, data) {
 
 // tab completion Symbol
 {
-  assert.strictEqual(typeof repl.replComplete, 'symbol');
+  assert.strictEqual(typeof repl.tabComplete, 'symbol');
 }
 
 // tab completion for function calls
@@ -216,14 +216,15 @@ testMe.complete('toSt', common.mustCall(function(error, data) {
   putIn.run(['var completeMethod = () => [7, 13]']);
 
   putIn.run(['var obj = { methodWithTabComplete: (i) => {} }']);
-  putIn.run(['obj.methodWithTabComplete[repl.replComplete] = completeMethod']);
+  putIn.run(['obj.methodWithTabComplete[repl.tabComplete] = completeMethod']);
 
   putIn.run(['var fnWithTabComplete = (i) => {}']);
-  putIn.run(['fnWithTabComplete[repl.replComplete] = completeMethod']);
+  putIn.run(['fnWithTabComplete[repl.tabComplete] = completeMethod']);
 
 
   ['obj.methodWithTabComplete(', 'fnWithTabComplete('].forEach(input => {
     testMe.complete(input, common.mustCall((err, data) => {
+      console.log(data)
       assert.strictEqual(err, null);
       assert.strictEqual(data.length, 2);
       assert.strictEqual(data[1], undefined);
@@ -233,11 +234,11 @@ testMe.complete('toSt', common.mustCall(function(error, data) {
     }));
   })
 
-  // does not tab complete expressions without the `replComplete` method
+  // does not tab complete expressions without the `tabComplete` method
   putIn.run(['.clear']);
   putIn.run(['var fnWithTabComplete = (i) => {} }']);
 
-  testMe.complete('fnWithTabComplete(', common.mustCall((err, data) => {
+  testMe.complete('fnWithTabComplete (', common.mustCall((err, data) => {
     assert.strictEqual(err, null);
     assert.strictEqual(data.length, 2);
     assert.strictEqual(data[1], undefined);
@@ -245,21 +246,46 @@ testMe.complete('toSt', common.mustCall(function(error, data) {
   }));
 
   // passes input params to the completer function
-  putIn.run(['.clear']);
-  putIn.run(['var repl = require(\'repl\')']);
+  {
+    putIn.run(['.clear']);
+    putIn.run(['var repl = require(\'repl\')']);
 
-  putIn.run(['var completeMethod = (args) => args']);
+    putIn.run(['var completeMethod = (args) => args']);
 
-  putIn.run(['var fnWithTabComplete = (i) => {}']);
-  putIn.run(['fnWithTabComplete[repl.replComplete] = completeMethod']);
+    putIn.run(['var fnWithTabComplete = (i) => {}']);
+    putIn.run(['fnWithTabComplete[repl.tabComplete] = completeMethod']);
+    const tabCompleteCases = [{
+      input: 'fnWithTabComplete (\'hi',
+      expected: ['\'hi'],
+    }, {
+      input: 'fnWithTabComplete (\'a\', \'b\',',
+      expected: ['\'a\'', '\'b\'', ''],
+    }, {
+      input: 'fnWithTabComplete(\"a\", b = 0, /*c,*/ d',
+      expected: ['\"a\"', 'b = 0', '/*c,*/ d']
+    }, {
+      input: 'fnWithTabComplete(a, \"bamieh, ahm\", \"incom::',
+      expected: ['a', '\"bamieh, ahm\"', '\"incom::']
+    }];
 
-  testMe.complete('fnWithTabComplete(\'hi', common.mustCall((err, data) => {
-    assert.strictEqual(err, null);
-    assert.strictEqual(data.length, 2);
-    assert.strictEqual(data[1], undefined);
-    assert.strictEqual(data[0].length, 1);
-    assert(data[0].includes('\'hi'));
-  }));
+    tabCompleteCases.forEach(tabCompleteCase => {
+      console.log("tabCompleteCase.input:", tabCompleteCase.input)
+      testMe.complete(tabCompleteCase.input, common.mustCall((err, data) => {
+        assert.strictEqual(err, null);
+        assert.strictEqual(data.length, 2);
+        assert.strictEqual(data[1], undefined);
+        console.log('data[0]', JSON.stringify(data[0], 2,2))
+        console.log('tabCompleteCase.expected', tabCompleteCase.expected)
+        console.log('data[0].length:', data[0].length)
+        console.log('tabCompleteCase.expected.length:', tabCompleteCase.expected.length)
+        assert.strictEqual(data[0].length, tabCompleteCase.expected.length);
+        tabCompleteCase.expected.forEach(expectedData => {
+          assert(data[0].includes(expectedData));
+        })
+      }));
+    })
+
+  }
 
 }
 // Tab complete provides built in libs for require()
